requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.5
django-filter==2.4.0
django-import-export==2.5.0
djangorestframework==3.12.2
djangorestframework-gis==0.17
et-xmlfile==1.0.1
idna==2.10
importlib-metadata==3.4.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==2.11.2
Markdown==3.3.3
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.6
psycopg2-binary==2.8.6
pytz==2020.5
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
tablib==3.0.0
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
